Data.Dimensions.Internal

Description

This module gathers and exports all parts of the units package that might be useful, even when going past the abstraction layer of the package.

With the exports from this module, it is possible to perform unsafe operations that do not respect the rules of dimensional analysis. Use with caution.

Additionally, no attempt will be made to keep the exports of this module backward compatible.

The Dim type

newtype Dim a

Dim adds a dimensional annotation to a Double. This is the representation for all dimensioned quantities.

Constructors

Dim Double

Manipulating dimension specifications

data DimSpec star

This will only be used at the kind level. It either holds a dimension with its exponent, or the special constant DAny, which can be any combination of dimensions at any exponents. It is used to represent multiplying by 0 somewhere.

Constructors

D star Z

DAny

type family a ($=) b :: Bool

Do these DimSpecs represent the same dimension?

type family Extract s lst :: ([DimSpec *], Maybe (DimSpec *))

(Extract s lst) pulls the DimSpec that matches s out of lst, returning a diminished list and, possibly, the extracted DimSpec.

 Extract A [A, B, C] ==> ([B, C], Just A
 Extract D [A, B, C] ==> ([A, B, C], Nothing)

type family Reorder a b :: [DimSpec *]

Reorders a to be the in the same order as b, putting entries not in b at the end

 Reorder [A 1, B 2] [B 5, A 2] ==> [B 2, A 1]
 Reorder [A 1, B 2, C 3] [C 2, A 8] ==> [C 3, A 1, B 2]
 Reorder [A 1, B 2] [B 4, C 1, A 9] ==> [B 2, A 1]
 Reorder x x ==> x
 Reorder x [] ==> x
 Reorder [] x ==> []

type family HasAny lst :: Bool

Check if a [DimSpec *] has a DAny inside it

type family a (@~) b :: Constraint

Check if two [DimSpec *]s should be considered to be equal

type family Normalize d :: [DimSpec *]

If a [DimSpec *] has a DAny, collapse the whole list to one DAny. Otherwise, normalize the list by removing exponents of 0.

type family ChooseFrom d1 d2 :: [DimSpec *]

Given two [DimSpec *]s, return the one that lacks a DAny, if there is one.

type family a (@+) b :: [DimSpec *]

Adds corresponding exponents in two dimension

type family a (@-) b :: [DimSpec *]

Subtract exponents in two dimensions

type family NegDim a :: DimSpec *

negate a single DimSpec

type family NegList a :: [DimSpec *]

negate a list of DimSpecs

type family base (@*) power :: [DimSpec *]

Multiplication of the exponents in a dimension by a scalar

type family dims (@/) z :: [DimSpec *]

Division of the exponents in a dimension by a scalar

Generally-useful type operations

type family Fst x :: a

Extract the first element of a pair

type family Snd x :: b

Extract the second element of a pair

type family If switch true false :: k

Type-level conditional

type family a :&&: b :: Bool

Type-level and

type family a :||: b :: Bool

Type-level or

type family a :=: b :: Bool

Type-level equality over *.