Data.Dimensions.SI

Description

This module exports unit, type, and prefix definitions according to the SI system of units. The definitions were taken from here: http://www.bipm.org/en/si/.

There is one deviation from the definition at that site: To work better with prefixes, the unit of mass is Gram.

Documentation

module Data.Dimensions.SI.Units

module Data.Dimensions.SI.Types

module Data.Dimensions.SI.Prefixes