Data.Dimensions.Unsafe

Description

This module exports the constructor of the Dim type. This allows you to write dimension-unsafe code. Use at your peril.

The Dim type

newtype Dim n a

Dim adds a dimensional annotation to its base type n. This is the representation for all dimensioned quantities.

Constructors

Dim n