Data.Dimensions.SI

Description

This module exports unit definitions according to the SI system of units. The definitions were taken from here: http://www.bipm.org/en/si/.

There is one deviation from the definition at that site: To work better with prefixes, the unit of mass is Gram.

Documentation

data Meter

Constructors

Meter

data Gram

Constructors

Gram

data Second

Constructors

Second

data Ampere

Constructors

Ampere

data Kelvin

Constructors

Kelvin

data Mole

Constructors

Mole

data Candela

Constructors

Candela

data Hertz

Constructors

Hertz

data Newton

Constructors

Newton

data Pascal

Constructors

Pascal

data Joule

Constructors

Joule

data Watt

Constructors

Watt

data Coulomb

Constructors

Coulomb

data Volt

Constructors

Volt

data Farad

Constructors

Farad

data Ohm

Constructors

Ohm

data Siemens

Constructors

Siemens

data Weber

Constructors

Weber

data Tesla

Constructors

Tesla

data Henry

Constructors

Henry

data Lumen

Constructors

Lumen

data Lux

Constructors

Lux

data Becquerel

Constructors

Becquerel

data Gray

Constructors

Gray

data Sievert

Constructors

Sievert

data Katal

Constructors

Katal